requirements.txt
gradio==3.50.2
pytesseract
pdf2image
Pillow
transformers
torch
